requirements.txt
fastapi==0.82.0
uvicorn[standard] == 0.18.3
SQLAlchemy==1.4.40
mysqlclient==2.1.1
alembic==1.8.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
requests==2.28.1